+++ src/core/camera.cpp
@@ -5,7 +5,7 @@
 
 Camera::Camera(Vector3D &origin, 
                Vector3D &view_point, 
-               double fov,
+               float fov,
                int height,
                int width)
 {
@@ -13,7 +13,7 @@ Camera::Camera(Vector3D &origin,
   auto origin_transformed = origin.to_blender();
 
   this->origin = origin_transformed;
-  this->aspect = double(width) / double(height);
+  this->aspect = float(width) / float(height);
   this->width = width;
   this->height = height;
   Vector3D worldUp(0,1,0);
@@ -30,12 +30,12 @@ Camera::~Camera()
 
 Vector3D Camera::pixelToWorldSpace(int i, int j) const
 {
-  double ndcX = (2*(i + 0.5)/width  - 1) * aspect * scale;
-  double ndcY = (1 - 2*(j + 0.5)/height) * scale;
+  float ndcX = (2*(i + 0.5)/width  - 1) * aspect * scale;
+  float ndcY = (1 - 2*(j + 0.5)/height) * scale;
   return ((right*ndcX + up*ndcY) + front).normalized();
 }
 
-void Camera::zoom(double z)
+void Camera::zoom(float z)
 {
     Ray r(origin, front);
     this->origin = this->origin + r.at(z);
